% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_exercise_results.R
\name{display_exercise_results}
\alias{display_exercise_results}
\title{Display the results from an evaluated exercise in the chunk}
\usage{
display_exercise_results(
  exercise,
  envir,
  evaluate_global_setup = FALSE,
  err = NULL
)
}
\description{
Display the results from an evaluated exercise in the chunk
}
