% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_exercise2.R
\name{evaluate_exercise2}
\alias{evaluate_exercise2}
\title{Danny's version of exercise evaluation}
\usage{
evaluate_exercise2(exercise, envir, evaluate_global_setup = FALSE)
}
\description{
Danny's version of exercise evaluation
}
