% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trivial_evaluator.R
\name{trivial_evaluator}
\alias{trivial_evaluator}
\title{Trivial exercise evaluator for demonstration purposes}
\usage{
trivial_evaluator(exercise, envir, evaluate_global_setup = FALSE)
}
\description{
Trivial exercise evaluator for demonstration purposes
}
